linux-firmware: Fix Bt and WiFi not working since v2.48.0
Starting with v2.48 meta-balena switched to another
recipe for linux-firmware. The variscite layer _git.bbappend
handling BT and WIFI thus no longer applies to the installed
package.

Bring in this recipe append in our layers so that we can rename
and apply it on top of the new firmware package recipe.

Signed-off-by: Alexandru Costache <alexandru@balena.io>