u-boot: Configure loading and starting of M4
A file named m4_fw.bin needs to be
placed in /mnt/boot and the M4 core
will be started by uboot, but its' presence
is not mandatory. TCM is used for loading
the binary in this case, to switch to DDR
the M4 dtb will need to be used.

Let's stay for now on the default DTB
as the M4 one disables spi.

Signed-off-by: Alexandru Costache <alexandru@balena.io>