linux-variscite: Disable CQE directly from the mmc subsystem
The previous patch to disable CQE was faulty because it was missing the
changes to take into account the no-cqe device tree property. We now
just disable CQE directly from the mmc subsystem. This also
has the advantage that we disable CQE regardless of what device tree we
may be using.

Changelog-entry: Correctly disable CQE for imx8mm-var-dart-nrt
Signed-off-by: Florin Sarbu <florin@balena.io>