resin-image: Don't include gpu vivante driver for imx8mm-var-dart-nrt
Since we updated to kernel version 4.19.35 we would need to also add the
vivante patches from the BSP layer to make this driver compile with the
newer kernel version. But since the customer wishes to disable the GPU
anyway on this board we just do not compile this driver at all for the
imx8mm-var-dart-nrt machine that the customer uses.

Signed-off-by: Florin Sarbu <florin@balena.io>