resin-image.inc: Adapt to new dtb name and disable vivante kernel driver
We need to adapt to the new dtb name in kernel 5.4.3.

Also we disable building kernel-module-imx-gpu-viv since customer is not
using accelerated graphics. In order to build this module we'd need to
pull in additional recipes from a newer BSP layer so that we are
compatible with the newer kernel we switched to.

Signed-off-by: Florin Sarbu <florin@balena.io>