linux-variscite: Add 4.19.35 for NRT and 5.4.85 recipes
The NRT version is added because it is no longer available
in upstream dunfell branch and uses rt patches that are
available only for this version. The 5.4.85 version is
added so we can set the version properly, otherwise
the mx8mm continues to use 5.4.3 wich is for the PLT
variant.

Signed-off-by: Alexandru Costache <alexandru@balena.io>