u-boot-variscite: Rework patches, add workaround for NRT uart
The NRT device appears to constantly reboot on
production images IF no uart is connected to
the first serial. This only happens on the NRT
device, and not with the NRT image on the devkit.

We now add the necessary u-boot configs
for only for production NRT images and
have a separate issue 2217 in upstream
meta-balena tracking to fix applying the
right configs for production uart.

We also remove the m4 patch since
it's not used currently because it
occupies one i2c bus and one uart.

Signed-off-by: Alexandru Costache <alexandru@balena.io>