u-boot-variscite: Remove usage of old dtb on altboot
The workaround in the previous commit doesn't appear to work
as the old 4.19.35 kernel hangs 2 seconds into booting,
way before the initramfs scripts are run, even if the corresponding
dtb is used.

So the U-Boot 2020.04-imx_v2020.04_5.4.24 will only boot
the kernel 5.4 succesfully.

Leaving this note so we know altboot doesn't work
when switching from kernel 4.19.x to 5.4.x, even if
the corresponding dtb version is used.

Signed-off-by: Alexandru Costache <alexandru@balena.io>