u-boot: Use old imx-boot on imx8mm-var-dart-plt during rollback
We need to use the corresponding imx-boot binary during
rollback-altboot if the kernel and device-tree  are
older than the current 5.4.24 imx-boot, otherwise the old kernel will
hang when booted with a newer imx-boot.

We thus write the old version of imx-boot during rollback-altboot
if the device-tree name in the old rootfs does not correspond to
the latest upstream naming scheme.

Changelog-entry: u-boot: Use old imx-boot on imx8mm-var-dart-plt during rollback
Signed-off-by: Alexandru Costache <alexandru@balena.io>