machines: do not set ROOTFS_SIZE
The machine configurations in meta-variscite-fslc all define
a fixed ROOTFS_SIZE, which should be autogenerated based on the actual rootfs
size.

balenaOS needs this variable to reflect the real size of the root
filesysem.

This commit brings a machine configuration used as include in other
machines as an include but removes the ROOTFS_SIZE setting.

Changelog-entry: do not set ROOTFS_SIZE in machine configurations
Signed-off-by: Alex Gonzalez <alexg@balena.io>