recipes-core/images: Don't install extra_uEnv.txt in flasher image
The extra_uEnv.txt file is installed by meta-balena in the non-flasher
image so that it can be manipulated by the supervisor, but there's
no other reason for including it in the flasher by the yocto build.

This was a leftover from testing the custom dtbs during provisioning,
let's remove it because it causes problems for the nodejs file
manipulation library during provisioning of boards which use a
button for toggling boot from sd-cards.

Changelog-entry: recipes-core/images: Don't install extra_uEnv.txt in flasher image
Signed-off-by: Alexandru Costache <alexandru@balena.io>